cloud-functions/upload-blob/requirements.txt
google-cloud-storage==2.*
coincurve==20.*
quart==0.19.*
hypercorn==0.17.*
